Review items:
Remove "run" from the error message that import is not supported;
Remove explicit passing in of nil in the CreateRunInput;
Fix handling of nil module version.

Also, remodel signatures of doApplyOrDestroyRun and getCurrentDeployment to fix error handling: return diag.Diagnostics from the functions; write the result we want through a pointer.  Otherwise, errors were dropping through the cracks in the function interface.