Remove the RunID field.
It was necessary to remove the plan-only run from the Read method.
The Update method will need to do a full (plan and apply) run whenever TF allows it.
Eventually, when the API/SDK support speculative plan with module source, that can be added.

Currently, when doing an apply with a change in variable value, TF errors out because has_changes has changed.
This is a checkpoint on the way to removing that field, as well.